Add the name of the plugin to the name of the generated function.
The getFunctionName function randomly generates a name to be assigned to the function responsible for executing the plugin. We add the name of the plugin to the function so that the plugin loaded as a module can look up the module name and find out its name. This allows for example a plugin to execute code based on its name.